--- src/main/java/edu/vanier/superspace/simulation/SimulationTimeCallback.java
@@ -1,6 +1,0 @@
-package edu.vanier.superspace.simulation;
-
-@FunctionalInterface
-public interface SimulationTimeCallback {
-    void update(double deltaTime);
-}
