+++ src/main/java/edu/vanier/superspace/simulation/Tickable.java
@@ -1,0 +1,5 @@
+package edu.vanier.superspace.simulation;
+
+public interface Tickable {
+    void onUpdate(double deltaTime);
+}
